1.新增JWT驗證 2..新增spring security驗證

+++ src/main/java/com/example/ballgame/repository/UsersRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.example.ballgame.dao.UserDao;
 
 @Repository
-public interface UsersRepository extends JpaRepository<UserDao,Integer>{
+public interface UsersRepository extends JpaRepository<UserDao,Long>{
 
 	Optional<UserDao> findByEmail(String email);
 }
